+++ examples/Example.BoundValues.cat/Program.cs
@@ -2,6 +2,7 @@
 namespace Example.BoundValues.cat
 {
     using System;
+    using System.Linq;
 
     class Program
     {
@@ -22,7 +23,7 @@ namespace Example.BoundValues.cat
             SqueezeEmptyAdjacentLines   =   0x00000010,
         }
 
-        [Clasp.Binding.BoundType]
+        [Clasp.Binding.BoundType(ParsingOptions=Clasp.ParseOptions.TreatSinglehyphenAsValue)]
         struct Arguments
         {
             /// <summary>
@@ -36,16 +37,20 @@ namespace Example.BoundValues.cat
             ///  Combination of <see cref="CatOptions"/> from CLI flags
             /// </summary>
             [Clasp.Binding.BoundEnumeration(typeof(CatOptions))]
-            [Clasp.Binding.BoundEnumerator("-n", (int)CatOptions.NumberLines)]
-            [Clasp.Binding.BoundEnumerator("-b", (int)CatOptions.NumberNonBlankLines)]
-            [Clasp.Binding.BoundEnumerator("-s", (int)CatOptions.SqueezeEmptyAdjacentLines)]
+            [Clasp.Binding.BoundEnumerator("--show-lines", (int)CatOptions.NumberLines, Alias="-n", HelpSection="display:", HelpDescription="shows the number of lines")]
+            [Clasp.Binding.BoundEnumerator("--show-blanks", (int)CatOptions.NumberNonBlankLines, Alias="-b", HelpSection="display:", HelpDescription="shows the number of non-blank lines")]
+            [Clasp.Binding.BoundEnumerator("--squeeze-empty", (int)CatOptions.SqueezeEmptyAdjacentLines, Alias="-s", HelpSection="display:", HelpDescription="squeeze adjacent empty lines into one")]
             public CatOptions Options;
+
+            [Clasp.Binding.BoundOption("--eol-sequence", Alias="-E", HelpSection="behaviour:", HelpDescription="specifies the end-of-line-sequence")]
+            public string EolSequence;
 
             #region implementation
             internal void _suppress_warning_CS0469()
             {
                 this.InputPaths = null;
                 this.Options = CatOptions.None;
+                this.EolSequence = null;
             }
             #endregion
         }
@@ -56,15 +61,21 @@ namespace Example.BoundValues.cat
             {
                 var specifications = new Clasp.Specification[] {
 
+                    Clasp.Specification.Section("display:"),
+
+                    Clasp.Specification.Section("behaviour:"),
+
+                    Clasp.Specification.Section("standard:"),
+
                     Clasp.Util.UsageUtil.Help,
                     Clasp.Util.UsageUtil.Version,
                 };
 
-                return Clasp.Invoker.ParseAndInvokeMain(argv, specifications, (Clasp.Arguments clargs) => {
+                return Clasp.Invoker.ParseAndInvokeMain<Program.Arguments>(argv, specifications, (Clasp.Arguments clargs) => {
 
                     if (clargs.HasFlag("--help"))
                     {
-                        return Clasp.Util.UsageUtil.ShowUsage(clargs, null);
+                        return Clasp.Util.UsageUtil.ShowBoundUsage<Program.Arguments>(clargs, null);
                     }
 
                     if (clargs.HasFlag("--version"))
@@ -74,7 +85,7 @@ namespace Example.BoundValues.cat
 
                     return Clasp.Invoker.InvokeMainWithBoundArgumentOfType<Program.Arguments>(clargs, (Program.Arguments prargs, Clasp.Arguments _) => {
 
-                        Console.Out.WriteLine("Behaviour (from flags):");
+                        Console.Out.WriteLine("Display (from flags):");
                         if(0 != (Program.CatOptions.NumberNonBlankLines & prargs.Options))
                         {
                             Console.Out.WriteLine("\twill number non-blank lines");
@@ -87,27 +98,32 @@ namespace Example.BoundValues.cat
                         {
                             Console.Out.WriteLine("\twill squeeze empty adjacent lines");
                         }
+                        Console.Out.WriteLine();
 
+                        Console.Out.WriteLine("Behaviour (from options):");
+                        if(!String.IsNullOrWhiteSpace(prargs.EolSequence))
+                        {
+                            Console.Out.WriteLine("\tEOL sequence: {0}", prargs.EolSequence.Select((ch) => (int)ch));
+                        }
+                        Console.Out.WriteLine();
 
                         Console.Out.WriteLine("Sources:");
                         foreach(string path in prargs.InputPaths)
                         {
                             if("-" == path)
                             {
-                                Console.Out.WriteLine("\tcource: would read from standard input stream");
+                                Console.Out.WriteLine("\tsource: would read from standard input stream");
                             }
                             else
                             {
-                                Console.Out.WriteLine("\tcource: would read from the path '{0}'", path);
+                                Console.Out.WriteLine("\tsource: would read from the path '{0}'", path);
                             }
                         }
+                        Console.Out.WriteLine();
 
                         return 0;
                     });
-                }
-                , Clasp.ParseOptions.TreatSinglehyphenAsValue
-                , Clasp.FailureOptions.Default
-                );
+                });
             }
             catch (System.OutOfMemoryException)
             {