+++ src/CLASP.Core/Binding/BoundFlagAttribute.cs
@@ -1,6 +1,6 @@
 ﻿
 // Created: 18th June 2017
-// Updated: 5th May 2019
+// Updated: 7th June 2019
 
 namespace Clasp.Binding
 {
@@ -29,6 +29,9 @@ namespace Clasp.Binding
             : base(flagName)
         {
         }
+        #endregion
+
+        #region properties
         #endregion
     }
 }