+++ src/CLASP.Core/Binding/BoundOptionAttribute.cs
@@ -1,6 +1,6 @@
 ﻿
 // Created: 18th June 2017
-// Updated: 5th May 2019
+// Updated: 29th May 2019
 
 namespace Clasp.Binding
 {